Several tweaks to the automated building of centrosome.

+++ package_for_windows.bat
@@ -7,80 +7,30 @@ REM Create release folder
 rmdir /s /q release
 mkdir release
 
-REM
-REM Python 3.9
-REM
+REM Loop through the Python versions using a for loop
+FOR %%V IN (3.9, 3.10, 3.11, 3.12) DO (
+    ECHO Building centrosome for Python %%V
 
-REM Create and activate a dedicated python 3.9 env
-call conda create -n centrosome-build python=3.9 -y
-call conda activate centrosome-build
+    REM Create and activate a dedicated Python environment
+    call conda create -n centrosome-build python=%%V -y
+    call conda activate centrosome-build
 
-REM Install dependencies
-poetry install
+    REM Install dependencies
+    call poetry install
 
-REM Delete build and dist folders
-rmdir /s /q build
-rmdir /s /q dist
+    REM Delete build and dist folders
+    IF EXIST build rmdir /s /q build
+    IF EXIST dist rmdir /s /q dist
 
-REM Build the binary package
-python setup.py bdist_wheel
+    REM Build the binary package
+    call python setup.py bdist_wheel
 
-REM Move it to release
-move /Y dist\*.whl release
+    REM Move it to release
+    IF NOT EXIST release mkdir release
+    move /Y dist\*.whl release\
 
-REM Remove the conda environment
-call conda deactivate
-call conda env remove -n centrosome-build
-IF EXIST "%ANACONDA_HOME%"\envs\centrosome-build rmdir /s /q "%ANACONDA_HOME%"\envs\centrosome-build
-
-REM
-REM Python 3.10
-REM
-
-REM Create and activate a dedicated python 3.10 env
-call conda create -n centrosome-build python=3.10 -y
-call conda activate centrosome-build
-
-REM Install dependencies
-poetry install
-
-REM Delete build and dist folders
-rmdir /s /q build
-rmdir /s /q dist
-
-REM Build the binary package
-python setup.py bdist_wheel
-
-REM Move it to release
-move /Y dist\*.whl release
-
-REM Remove the conda environment
-call conda deactivate
-call conda env remove -n centrosome-build
-IF EXIST "%ANACONDA_HOME%"\envs\centrosome-build rmdir /s /q "%ANACONDA_HOME%"\envs\centrosome-build
-
-REM
-REM Python 3.11
-REM
-
-REM Create and activate a dedicated python 3.11 env
-call conda create -n centrosome-build python=3.11 -y
-call conda activate centrosome-build
-
-REM Install dependencies
-poetry install
-
-REM Delete build and dist folders
-rmdir /s /q build
-rmdir /s /q dist
-
-REM Build the binary package
-python setup.py bdist_wheel
-
-REM Move it to release
-move /Y dist\*.whl release
-
-REM Remove the conda environment
-call conda deactivate
-call conda env remove -n centrosome-build
-IF EXIST "%ANACONDA_HOME%"\envs\centrosome-build rmdir /s /q "%ANACONDA_HOME%"\envs\centrosome-build
+    REM Remove the conda environment
+    call conda deactivate
+    call conda env remove -n centrosome-build -y
+    IF EXIST "%ANACONDA_HOME%\envs\centrosome-build" rmdir /s /q "%ANACONDA_HOME%\envs\centrosome-build"
+)
